Add a new `wit-smith` subcommand
This commit extracts the WIT document generation from the
`roundtrip-wit` fuzzer into a dedicated crate and a dedicated
subcommand, `wasm-tools wit-smith`. This is intended to behave similarly
to the `wasm-tools smith` subcommand but it instead generates a
WebAssembly binary-encoded WIT document.
