+++ .github/workflows/main.yml
@@ -89,6 +89,7 @@ jobs:
       - run: cargo check --no-default-features --features demangle
       - run: cargo check --no-default-features --features component
       - run: cargo check --no-default-features --features metadata
+      - run: cargo check --no-default-features --features wit-smith
 
   doc:
     runs-on: ubuntu-latest
