+++ Cargo.toml
@@ -54,6 +54,7 @@ wast = { version = "56.0.0", path = "crates/wast" }
 wat = { version = "1.0.62", path = "crates/wat" }
 wit-component = { version = "0.8.1", path = "crates/wit-component" }
 wit-parser = { version = "0.7.0", path = "crates/wit-parser" }
+wit-smith = { version = "0.1.0", path = "crates/wit-smith" }
 
 [dependencies]
 anyhow = { workspace = true }
@@ -103,6 +104,9 @@ wast = { workspace = true, optional = true }
 # Dependencies of `metadata`
 wasm-metadata = { workspace = true, features = ["clap"], optional = true }
 
+# Dependencies of `wit-smith`
+wit-smith = { workspace = true, features = ["clap"], optional = true }
+
 [dev-dependencies]
 serde_json = "1.0"
 tempfile = "3.1"
@@ -134,6 +138,7 @@ default = [
   'demangle',
   'component',
   'metadata',
+  'wit-smith',
 ]
 
 # Each subcommand is gated behind a feature and lists the dependencies it needs
@@ -150,3 +155,4 @@ compose = ['wasm-compose']
 demangle = ['rustc-demangle', 'cpp_demangle', 'wasmparser', 'wasm-encoder']
 component = ['wit-component', 'wit-parser', 'wast', 'wasm-encoder', 'wasmparser']
 metadata = ['wasmparser', 'wasm-metadata', 'serde_json' ]
+wit-smith = ['dep:wit-smith']
