+++ crates/wit-smith/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+description = "A WIT test case generator"
+documentation = "https://docs.rs/wit-smith"
+edition.workspace = true
+license = "Apache-2.0 WITH LLVM-exception"
+name = "wit-smith"
+repository = "https://github.com/bytecodealliance/wasm-tools/tree/main/crates/wit-smith"
+version = "0.1.0"
+
+[dependencies]
+arbitrary = { workspace = true, features = ["derive"] }
+wit-component = { workspace = true }
+wit-parser = { workspace = true }
+clap = { workspace = true, optional = true }
